pyproject.toml
[project]
name = "sidekick"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "gradio>=5.44.1",
    "ipython>=9.5.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "nest-asyncio>=1.6.0",
    "playwright>=1.55.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.11",
]
